Add Virtual Try-On page with navigation
